查看近期考试

+++ src/main/java/com/czy/template/view/dto/PublicExamDTO.java
@@ -9,7 +9,7 @@ public class PublicExamDTO {
     private Long examId;                   //考试id
     private String title;                 //考试名称
     private List<Integer> questionIds;    //考试题目
-    private int totalScore;               //总分数
-    private int totalTitle;               //总题目数
+    private Double totalScore;               //总分数
+    private Integer totalTitle;               //总题目数
     private String exam_uuid;             //考试编号
 }
